Break into statustype and internal/status
