+++ security/advancedtls/go.mod
@@ -2,4 +2,9 @@ module google.golang.org/grpc/security/advancedtls
 
 go 1.13
 
-require google.golang.org/grpc v1.27.0
+require (
+	golang.org/x/lint v0.0.0-20190313153728-d0100b6bd8b3 // indirect
+	golang.org/x/tools v0.0.0-20190524140312-2c0ae7006135 // indirect
+	google.golang.org/grpc v1.27.0
+	honnef.co/go/tools v0.0.0-20190523083050-ea95bdfd59fc // indirect
+)
