+++ status/status_test.go
@@ -24,6 +24,10 @@ import (
 	"fmt"
 	"testing"
 
+	"google.golang.org/grpc/internal/status/statustype"
+
+	"google.golang.org/grpc/internal/status"
+
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes"
 	apb "github.com/golang/protobuf/ptypes/any"
@@ -34,7 +38,6 @@ import (
 	spb "google.golang.org/genproto/googleapis/rpc/status"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/internal/grpctest"
-	"google.golang.org/grpc/internal/status"
 )
 
 type s struct {
@@ -154,12 +157,18 @@ func (c customError) Error() string {
 	return fmt.Sprintf("rpc error: code = %s desc = %s", c.Code, c.Message)
 }
 
-func (c customError) GRPCStatus() *spb.Status {
-	return &spb.Status{
-		Code:    int32(c.Code),
-		Message: c.Message,
-		Details: c.Details,
+func (c customError) GRPCStatus() *Status {
+	s := statustype.New(c.Code, c.Message)
+	var a []proto.Message
+	for _, d := range c.Details {
+		any, err := ptypes.MarshalAny(d)
+		if err != nil {
+			break
+		}
+		a = append(a, any)
 	}
+	s, _ = s.WithDetails(a...)
+	return s
 }
 
 func (s) TestFromErrorImplementsInterface(t *testing.T) {
@@ -343,7 +352,7 @@ func str(s *Status) string {
 	if s.s == nil {
 		return "<Code=OK>"
 	}
-	return fmt.Sprintf("<Code=%v, Message=%q, Details=%+v>", codes.Code(s.s.GetCode()), s.s.GetMessage(), s.s.GetDetails())
+	return fmt.Sprintf("<Code=%v, Message=%q, Details=%+v>", s.Code(), s.Message(), s.Details())
 }
 
 // mustMarshalAny converts a protobuf message to an any.
